mcp_new_project/config/requirements/requirements.txt
# Core Dependencies
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0
uvicorn[standard]==0.27.0  # Updated from 0.24.0.post1

# Database - Async Support
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0  # PostgreSQL async driver
psycopg2-binary==2.9.9  # PostgreSQL driver for sync operations
aiosqlite==0.19.0  # SQLite async driver (development)
greenlet==3.0.1  # Required for SQLAlchemy async

# Async Utilities
aiofiles==23.2.1

# Background Tasks
celery==5.3.6  # Updated from 5.3.4
redis==5.0.1
# flower==2.0.1  # Commented out - unused dependency

# Real-time Communication
flask-socketio==5.3.6
python-socketio==5.11.1  # Updated from 5.11.0
# Replace eventlet with gevent due to security vulnerability
gevent==23.9.1  # Replaced eventlet==0.33.3
gevent-websocket==0.10.1

# Memory Integration
httpx==0.25.2  # For Supermemory API calls

# AI/ML Libraries - CRITICAL UPDATES
openai==1.35.0  # Updated from 1.3.5 - major security fixes
anthropic==0.25.0  # Updated from 0.7.8 - important updates

# Utilities
colorama==0.4.6  # Colored terminal output
python-dateutil==2.8.2
pytz==2023.3
pydantic==2.5.2

# Configuration & Validation
jsonschema==4.20.0  # For config validation
pyyaml==6.0.1  # For YAML support
marshmallow==3.20.1

# Retry & Resilience
tenacity==8.2.3
# apprise==1.6.0  # Commented out - unused dependency

# Observability & Monitoring
loguru==0.7.2
# sentry-sdk[flask,celery]==1.39.1  # Commented out - imported but not configured

# Security Enhancements
python-jose[cryptography]==3.3.0  # For JWT handling
cryptography==41.0.7  # For secure operations

# Development Tools - Move to requirements-dev.txt
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# flake8==6.1.0
# pre-commit==3.5.0

# Optional Performance Enhancements
# ujson==5.9.0  # Faster JSON parsing
# orjson==3.9.10  # Even faster JSON parsing
# msgpack==1.0.7  # Binary serialization